Move YAML::convert<Condition> into Parse.cpp for general use
