+++ parse/Parse.cpp
@@ -543,3 +543,39 @@ namespace parse {
         return parser_success && parse_length_good;
     }
 }}
+
+
+namespace YAML {
+bool convert<std::unique_ptr<Condition::Condition>>::decode(const Node& node, std::unique_ptr<Condition::Condition>& rhs) {
+    if (!node.IsScalar())
+        return false;
+
+    std::string data{node.as<std::string>()};
+
+    const parse::lexer lexer;
+    parse::detail::Labeller label;
+    parse::conditions_parser_grammar condition_parser(lexer, label);
+    boost::spirit::qi::in_state_type in_state;
+    parse::text_iterator begin = data.begin();
+    parse::text_iterator end = data.end();
+    parse::token_iterator it = lexer.begin(begin, end);
+
+    parse::detail::MovableEnvelope<Condition::Condition> location;
+
+    bool success = boost::spirit::qi::phrase_parse(
+        it, lexer.end(),
+        condition_parser, in_state("WS")[lexer.self],
+        location);
+
+    if (!success) {
+        ErrorLogger() << "monster_fleet_plans: failed to parse location attribute";
+        return false;
+    }
+
+    bool passed{true};
+
+    rhs = location.OpenEnvelope(passed);
+
+    return passed;
+}
+}
